Fixes max_width parameter
The max_width parameter was acting as a "set to width" parameter, meaning the canvas was being scaled to the given dimension even if the the image was smaller than the max_width. max_width should only affect gifs that are larger than the given max_width value. 

This adds an additional test to the get_canvas_scale method, checking if hdr.width is greater than the given max_width option before setting the canvas_scale to something other than 1.